+++ .github/workflows/build.yaml
@@ -1,0 +1,34 @@
+name: Build
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+
+concurrency:
+  group: ${{ github.workflow }}
+
+env:
+  ARTIFACT: ${{ runner.temp }}/python.vim
+
+jobs:
+  main:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+
+      - name: Setup
+        uses: actions/setup-python@v6
+        with:
+          python-version: '3.14'
+
+      - name: Run
+        run: |
+          python3 src/python2vim.py > "${{ env.ARTIFACT }}"
+
+      - name: Upload
+        uses: actions/upload-artifact@v4
+        with:
+          path: ${{ env.ARTIFACT }}
